make compatible with Silverstripe 4.0
- pending fix for "Call to a member function setFieldFormatting() on null" in MemeberProfilePage line 196
- pending fix for removed DataModel in code/formfields/MemberProfilesAddSectionAction.php line 81
