+++ code/controllers/MemberProfilesAddSectionAction_ItemRequest.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Silverstripe\MemberProfiles;
+use SilverStripe\Control\Controller;
+use SilverStripe\Forms\GridField\GridFieldDetailForm_ItemRequest;
+
+class MemberProfilesAddSectionAction_ItemRequest extends GridFieldDetailForm_ItemRequest {
+
+    public function Link($action = null) {
+        if($this->record->ID) {
+            return parent::Link($action);
+        } else {
+            return Controller::join_links(
+                $this->gridField->Link(), 'addsection', get_class($this->record)
+            );
+        }
+    }
+
+}
